Update README to record the death of this utility
*************** THIS PROGRAM IS NO LONGER MAINTAINED BY X.ORG ***************

xsetpointer sets an input device as the core pointer. This dates back to
when an extension device could not be a core device at the same time and
is obsolete since server 1.4 (released September 2007).
The request always fails now (96e32805d12fc36f0fa0926dbfb0dd8a5cadb739).

xinput set-pointer is the equivalent of xsetpointer, but still doesn't
do anything, this app is truly dead.

*****************************************************************************

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>